Refactor perf tests into individual areas
