+++ tests/performance/performance.vcxproj.filters
@@ -22,10 +22,10 @@
     <ClCompile Include="performance.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\end_to_end\test_helper.cpp">
+    <ClCompile Include="platform.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\libs\thunk\mock\mock.cpp">
+    <ClCompile Include="ExecutionContext.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
@@ -36,10 +36,7 @@
     <ClInclude Include="performance_measure.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\end_to_end\test_helper.hpp">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="..\end_to_end\mock.h">
+    <ClInclude Include="performance.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
